refactor get_input
