+++ crates/cli/src/input/get_input/data_admin_input/edit_data/mod.rs
@@ -1,0 +1,3 @@
+mod edit_data_input;
+
+pub use edit_data_input::*;
