+++ crates/cli/src/input/get_input/data_admin_input/mod.rs
@@ -1,0 +1,6 @@
+#[allow(clippy::module_inception)]
+mod data_admin_input;
+mod edit_data;
+
+pub use data_admin_input::*;
+pub use edit_data::*;
