+++ crates/cli/src/input/get_input/email_input/mod.rs
@@ -1,0 +1,6 @@
+mod edit_email_input;
+#[allow(clippy::module_inception)]
+mod email_input;
+
+pub use edit_email_input::*;
+pub use email_input::*;
