+++ crates/cli/src/input/get_input/expenses_input/mod.rs
@@ -1,0 +1,4 @@
+#[allow(clippy::module_inception)]
+mod expenses_input;
+
+pub use expenses_input::*;
