+++ crates/cli/src/input/get_input/get_input.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
 
-use clap::{Args, Parser, ValueEnum};
+use clap::{Args, Parser};
 use derive_more::{Debug, Unwrap};
 
 /// The root argument for the CLI, which contains the subcommands for
@@ -28,131 +28,6 @@ pub enum Command {
     Data(DataAdminInput),
 }
 
-#[derive(Debug, Args, Getters, PartialEq)]
-pub struct EmailInput {
-    #[command(subcommand)]
-    #[getset(get = "pub")]
-    command: EmailInputCommand,
-}
-
-#[derive(Debug, Subcommand, Unwrap, PartialEq)]
-pub enum EmailInputCommand {
-    /// Initializes the data related to sending emails in the data directory,
-    Init,
-    /// Validates the data related to sending emails in the data directory,
-    Validate,
-    Edit(EditEmailInput),
-    /// Sends an email with a sample invoice as PDF attachment using the data
-    /// in the data directory, which includes email account, SMTP server and
-    /// recipient information.
-    Test,
-}
-
-/// The CLI arguments for data management, such as initializing the data directory,
-/// validating the data, or recording expenses or month off.
-#[derive(Debug, Args, Getters, PartialEq)]
-pub struct DataAdminInput {
-    /// The command to run for data management, such as initializing the data directory,
-    /// validating the data, or recording expenses or month off.
-    #[command(subcommand)]
-    #[getset(get = "pub")]
-    command: DataAdminInputCommand,
-}
-
-/// The commands available for data management, such as initializing the data directory,
-/// validating the data, or recording expenses or month off.
-#[derive(Debug, Subcommand, Unwrap, PartialEq)]
-pub enum DataAdminInputCommand {
-    /// Prints the data in the data directory as a RON object.
-    Dump,
-    /// Initializes the data in the data directory, creating it if it does not exist.
-    /// Such as information about you as a vendor and your client, payment information
-    /// pricing etc
-    Init,
-    /// Validates the data in the data directory, checking if it is correctly formatted
-    /// and if all required fields are present.
-    Validate,
-    /// Just like `Init` but will use the existing data, prefilling the values
-    /// with the existing data as default values so that user can press Enter
-    /// to accept the existing values as defaults.
-    Edit(EditDataInput),
-    /// Records a month off for the specified month, which is used to calculate the invoice.
-    MonthOff(MonthOffInput),
-    /// Records expenses for the specified month, used to create expenses invoices
-    /// and affects invoice number calculation.
-    Expenses(ExpensesInput),
-}
-
-#[derive(Debug, Args, Getters, PartialEq)]
-pub struct EditEmailInput {
-    #[arg(value_enum)]
-    #[getset(get = "pub")]
-    selector: EditEmailInputSelector,
-}
-
-#[derive(Debug, Args, Getters, PartialEq)]
-pub struct EditDataInput {
-    #[arg(value_enum)]
-    #[getset(get = "pub")]
-    selector: EditDataInputSelector,
-}
-
-#[derive(Clone, Copy, Debug, Subcommand, Unwrap, PartialEq, ValueEnum)]
-#[clap(rename_all = "kebab_case")]
-pub enum EditDataInputSelector {
-    All,
-    Vendor,
-    Client,
-    Information,
-    PaymentInfo,
-    ServiceFees,
-}
-
-#[derive(Clone, Copy, Debug, Subcommand, Unwrap, PartialEq, ValueEnum)]
-#[clap(rename_all = "kebab_case")]
-pub enum EditEmailInputSelector {
-    All,
-    AppPassword,
-    EncryptionPassword,
-    Template,
-    Smtp,
-    ReplyTo,
-    Sender,
-    Recipients,
-    Cc,
-    Bcc,
-}
-
-impl From<EditEmailInputSelector> for EmailSettingsSelector {
-    fn from(selector: EditEmailInputSelector) -> Self {
-        match selector {
-            EditEmailInputSelector::All => EmailSettingsSelector::All,
-            EditEmailInputSelector::AppPassword => EmailSettingsSelector::AppPassword,
-            EditEmailInputSelector::EncryptionPassword => EmailSettingsSelector::EncryptionPassword,
-            EditEmailInputSelector::Template => EmailSettingsSelector::Template,
-            EditEmailInputSelector::Smtp => EmailSettingsSelector::SmtpServer,
-            EditEmailInputSelector::ReplyTo => EmailSettingsSelector::ReplyTo,
-            EditEmailInputSelector::Sender => EmailSettingsSelector::Sender,
-            EditEmailInputSelector::Recipients => EmailSettingsSelector::Recipients,
-            EditEmailInputSelector::Cc => EmailSettingsSelector::CcRecipients,
-            EditEmailInputSelector::Bcc => EmailSettingsSelector::BccRecipients,
-        }
-    }
-}
-
-impl From<EditDataInputSelector> for DataSelector {
-    fn from(selector: EditDataInputSelector) -> Self {
-        match selector {
-            EditDataInputSelector::All => DataSelector::All,
-            EditDataInputSelector::Vendor => DataSelector::Vendor,
-            EditDataInputSelector::Client => DataSelector::Client,
-            EditDataInputSelector::Information => DataSelector::Information,
-            EditDataInputSelector::PaymentInfo => DataSelector::PaymentInfo,
-            EditDataInputSelector::ServiceFees => DataSelector::ServiceFees,
-        }
-    }
-}
-
 /// Record a new month off for the specified month.
 #[derive(Debug, Args, Getters, PartialEq)]
 pub struct MonthOffInput {
@@ -165,30 +40,6 @@ pub struct MonthOffInput {
     )]
     #[getset(get = "pub")]
     month: YearAndMonth,
-}
-
-/// Record expenses for the specified month, which will be used to create expenses invoices
-/// and affects invoice number calculation.
-#[derive(Debug, Args, Getters, PartialEq)]
-pub struct ExpensesInput {
-    /// The month for which the expenses are recorded.
-    #[arg(
-        long,
-        short = 'm',
-        default_value = None,
-        help = "The month and year for which you wanna record expenses, e.g. `2025-05`. Note that we might expense for month of May even thought we had an expense in beginning of June, so this is not a strict month, but rather a month in which we want to record the expenses."
-    )]
-    #[getset(get = "pub")]
-    month: YearAndMonth,
-
-    /// The expenses to record for the month, which are specified as a list of items.
-    /// Please note that the transaction date might be different from the month specified,
-    /// so you can record expenses for a month even if the transaction date is in the next
-    /// month, e.g. you can record expenses for May even if the transaction date is in June.
-    /// Format for each item is: `name,amount,currency,quantity,date`, e.g. `Coffee,2.5,EUR,3.0,2025-05-31`.
-    #[arg(long, short = 'e', help = "The expenses to record for the month.")]
-    #[getset(get = "pub")]
-    expenses: Vec<Item>,
 }
 
 /// The CLI arguments for generating an invoice PDF.
@@ -293,23 +144,13 @@ mod tests {
         #[test]
         fn test_data_admin_init() {
             let input = CliArgs::parse_from([BINARY_NAME, "data", "init"]);
-            assert!(matches!(
-                input.command,
-                Command::Data(DataAdminInput {
-                    command: DataAdminInputCommand::Init
-                })
-            ));
+            assert!(matches!(input.command, Command::Data(_)));
         }
 
         #[test]
         fn test_data_admin_validate() {
             let input = CliArgs::parse_from([BINARY_NAME, "data", "validate"]);
-            assert!(matches!(
-                input.command,
-                Command::Data(DataAdminInput {
-                    command: DataAdminInputCommand::Validate
-                })
-            ));
+            assert!(matches!(input.command, Command::Data(_)));
         }
 
         #[test]
@@ -331,10 +172,12 @@ mod tests {
             ]);
             assert_eq!(
                 *input.command.unwrap_data().command(),
-                DataAdminInputCommand::Expenses(ExpensesInput {
-                    month: YearAndMonth::from_str("2025-05").unwrap(),
-                    expenses: vec![item_1, item_2]
-                })
+                DataAdminInputCommand::Expenses(
+                    ExpensesInput::builder()
+                        .month(YearAndMonth::from_str("2025-05").unwrap())
+                        .expenses(vec![item_1, item_2])
+                        .build()
+                )
             );
         }
     }
