+++ crates/cli/src/input/get_input/mod.rs
@@ -1,0 +1,10 @@
+mod data_admin_input;
+mod email_input;
+mod expenses_input;
+#[allow(clippy::module_inception)]
+mod get_input;
+
+pub use data_admin_input::*;
+pub use email_input::*;
+pub use expenses_input::*;
+pub use get_input::*;
