+++ crates/core/src/logic/command/mod.rs
@@ -1,0 +1,10 @@
+#[allow(clippy::module_inception)]
+mod command;
+mod data_selector;
+mod email_settings_selector;
+mod select;
+
+pub use command::*;
+pub use data_selector::*;
+pub use email_settings_selector::*;
+pub use select::*;
