+++ crates/core/src/logic/command/select.rs
@@ -1,0 +1,3 @@
+pub trait Select {
+    fn includes(&self, target: Self) -> bool;
+}
